+++ src/main/java/zb/accountMangement/account/controller/TransactionController.java
@@ -8,7 +8,7 @@ import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import zb.accountMangement.account.dto.TransactionDto;
 import zb.accountMangement.account.service.TransactionService;
-import zb.accountMangement.common.util.ValidationService;
+import zb.accountMangement.common.service.ValidationService;
 
 import javax.validation.Valid;
 import javax.validation.constraints.Min;
